+++ src/js/page_sections/Menu/SubmenuItem.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import I18nNavLink from 'js/modules/I18nNavLink';
+
+
+const SubmenuItem = ({ id, topic, handleClick }) => (
+  <li key={id} onClick={handleClick} className="coa-Menu__subitem"
+    role="menuitem"
+  >
+    <I18nNavLink to={`/topics/${topic.slug}`}
+      activeClassName="usa-current"
+      className="test"
+      tabIndex={-1}
+    >
+      {topic.title}
+    </I18nNavLink>
+  </li>
+)
+
+export default SubmenuItem;
